requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
WTForms==3.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
email-validator==2.0.0
flask-migrate==4.1.0
Pillow==10.0.0